+++ pennylane_lightning/lightning_kokkos/_adjoint_jacobian.py
@@ -20,7 +20,7 @@ from __future__ import annotations
 from warnings import warn
 
 try:
-    from pennylane_lightning.lightning_kokkos_ops.algorithms import (
+    from pennylane_lightning.lightning_kokkos_nb.algorithms import (
         AdjointJacobianC64,
         AdjointJacobianC128,
         create_ops_listC64,
@@ -28,7 +28,7 @@ try:
     )
 
     try:
-        from pennylane_lightning.lightning_kokkos_ops.algorithmsMPI import (
+        from pennylane_lightning.lightning_kokkos_nb.algorithmsMPI import (
             AdjointJacobianMPIC64,
             AdjointJacobianMPIC128,
             create_ops_listMPIC64,
@@ -40,7 +40,6 @@ try:
     except ImportError as ex_mpi:
         mpi_error = ex_mpi
         MPI_SUPPORT = False
-
 
 except ImportError as ex:
     warn(str(ex), UserWarning)