+++ pennylane_lightning/lightning_kokkos/_state_vector.py
@@ -17,7 +17,7 @@ Class implementation for lightning_kokkos state-vector manipulation.
 from warnings import warn
 
 try:
-    from pennylane_lightning.lightning_kokkos_ops import (
+    from pennylane_lightning.lightning_kokkos_nb import (
         InitializationSettings,
         StateVectorC64,
         StateVectorC128,
@@ -26,7 +26,7 @@ try:
     )
 
     try:
-        from pennylane_lightning.lightning_kokkos_ops import (
+        from pennylane_lightning.lightning_kokkos_nb import (
             MPIManagerKokkos,
             StateVectorMPIC64,
             StateVectorMPIC128,
@@ -216,9 +216,13 @@ class LightningKokkosStateVector(LightningBaseStateVector):
             state = state.toarray().flatten()
 
         if isinstance(state, self._qubit_state.__class__):
-            state_data = allocate_aligned_array(state.size, np.dtype(self.dtype), True)
+            state_data = allocate_aligned_array(state.size(), np.dtype(self.dtype), True)
             state.DeviceToHost(state_data)
             state = state_data
+
+        # Convert PennyLane tensor to NumPy array if needed
+        if hasattr(state, "numpy"):
+            state = state.numpy()
 
         if len(device_wires) == self._num_wires and Wires(sorted(device_wires)) == device_wires:
             # Initialize the entire device state with the input state